refactored Transaction view so that it can be used elsewhere moved the formatting methods in to formatting file removed unnecessary objects in User pages
