+++ lib/view/pages/user_page.dart
@@ -1,12 +1,10 @@
 import 'package:budgetflow/model/budget/budget_category.dart';
-import 'package:budgetflow/model/budget/transaction/transaction.dart';
-import 'package:budgetflow/model/budget/transaction/transaction_list.dart';
 import 'package:budgetflow/view/budgeting_app.dart';
 import 'package:budgetflow/view/global_widgets/main_drawer.dart';
+import 'package:budgetflow/view/global_widgets/transaction_view.dart';
 import 'package:budgetflow/view/utils/output_formater.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:intl/intl.dart';
 import 'package:pie_chart/pie_chart.dart';
 
 import 'add_transaction.dart';
@@ -17,13 +15,6 @@ class UserPage extends StatefulWidget {
 }
 
 class _UserPageState extends State<UserPage> {
-  Future _load;
-
-  @override
-  void initState() {
-    super.initState();
-    _load = BudgetingApp.userController.initialize();
-  }
 
   Widget _initUserPage() {
     return Scaffold(
@@ -41,12 +32,9 @@ class _UserPageState extends State<UserPage> {
                 showChartValues: true,
                 showLegends: true,
                 colorList: Colors.primaries,
-                showChartValuesOutside: true,
-                showChartValueLabel: true,
                 chartType: ChartType.ring,
               )),
           Card(
-            /*user cash flow*/
               child: RichText(
                   textAlign: TextAlign.center,
                   text: TextSpan(children: <TextSpan>[
@@ -104,15 +92,17 @@ class _UserPageState extends State<UserPage> {
                 },
               )
           ),
-          new _TransactionListView()
+          new TransactionListView()
         ],
       ),
       floatingActionButton: FloatingActionButton(
+        tooltip: 'add transaction',
         child: Icon(Icons.add),
         onPressed: () {
           Navigator.pushNamed(context, AddTransaction.ROUTE);
         },
       ),
+      floatingActionButtonLocation: FloatingActionButtonLocation.endFloat,
     );
   }
 
@@ -120,7 +110,7 @@ class _UserPageState extends State<UserPage> {
   @override
   Widget build(BuildContext context) {
     return FutureBuilder(
-        future: _load,
+        future: BudgetingApp.userController.initialize(),
         builder: (context, snapshot) {
           if (snapshot.hasData) {
             return _initUserPage();
@@ -131,74 +121,4 @@ class _UserPageState extends State<UserPage> {
   }
 } // _UserPage
 
-class _TransactionListView extends StatelessWidget {
-  final double _topRowFontSize = 20;
-  final double _bottomRowFontSize = 16;
 
-  @override
-  Widget build(BuildContext context) {
-    TransactionList transactions =
-    BudgetingApp.userController.getLoadedTransactions();
-    print('Generating ' +
-        transactions.length.toString() +
-        ' transaction list items');
-    return Card(
-        child: new ListView.builder(
-            shrinkWrap: true,
-            padding: EdgeInsets.all(8),
-            scrollDirection: Axis.vertical,
-            physics: ScrollPhysics(),
-            itemCount: transactions.length,
-            itemBuilder: (context, index) {
-              return _buildTransactionListViewItem(transactions.getAt(index));
-            }));
-  }
-
-  Widget _buildTransactionListViewItem(Transaction t){
-    return Table(children: [
-      TableRow(children: [
-        Text(
-          t.vendor,
-          textAlign: TextAlign.left,
-          style: TextStyle(fontSize: _topRowFontSize),
-        ),
-        Text(_formatDelta(t.delta),
-            textAlign: TextAlign.right,
-            style: TextStyle(
-                fontSize: _topRowFontSize, color: _deltaColor(t.delta)))
-      ]),
-      TableRow(children: [
-        Text(_formatDate(t.datetime),
-            textAlign: TextAlign.left,
-            style: TextStyle(fontSize: _bottomRowFontSize)),
-        Text(categoryJson[t.category],
-            textAlign: TextAlign.right,
-            style: TextStyle(fontSize: _bottomRowFontSize))
-      ])
-    ]);
-  }
-
-  String _formatDelta(double delta) {
-    String output = delta.toStringAsFixed(2);
-    if (output.contains('-')) {
-      output = output.replaceAll('-', '-\$');
-    } else {
-      output = '\$' + output;
-    }
-    return output;
-  }
-
-  MaterialColor _deltaColor(double delta) {
-    if (delta < 0) {
-      return Colors.red;
-    } else {
-      return Colors.green;
-    }
-  }
-
-  String _formatDate(DateTime dateTime) {
-    DateFormat dMy = new DateFormat('LLLL d, y');
-    DateFormat jm = new DateFormat('jm');
-    return dMy.format(dateTime) + ' ' + jm.format(dateTime);
-  }
-}
